getBinaryError

For potential k values, generate a binary error matrix from KNN label classification

Usage

getBinaryError(knn, k_values, class_train, class_true)

Arguments

knn

Is a k-nearest neighbour matrix, giving the indices of the training set that the query is closest to. Rows are the query cells, columns are the NNs, should be a large value. Typically output using BiocNeighbors::queryKNN(,,k = max(k_values)).

k_values

Is an integer vector of the values of k to consider for extracting accuracy. If k_values has names then pass these to colnames of E.

class_train

Is a factor or character vector of classes that corresponds to the indices given within knn.

class_true

Is a factor or character vector that corresponds to the rows of knn. If class_true has names then pass these to rownames of E.

Value

A sparse binary error matrix.